test(key-gen): move integration tests
