+++ oprf-key-gen/Cargo.toml
@@ -78,8 +78,9 @@ zeroize = { workspace = true, features = ["derive"] }
 [dev-dependencies]
 alloy = { workspace = true, features = ["node-bindings"] }
 axum-test = { workspace = true, features = ["ws"] }
-oprf-test-utils = { package = "taceo-oprf-test-utils", path = "../oprf-test-utils", version = "0.14" }
+nodes-common = { workspace = true, features = ["test-utils", "web3-asserter"] }
 parking_lot = { workspace = true }
+serde_json = { workspace = true }
 
 [target.'cfg(not(target_env = "msvc"))'.dependencies]
 tikv-jemallocator = { workspace = true }
