+++ oprf-key-gen/src/postgres/tests.rs
@@ -14,7 +14,6 @@ use groth16_material::circom::{CircomGroth16Material, CircomGroth16MaterialBuild
 use nodes_common::postgres::{PostgresConfig, SanitizedSchema};
 use nodes_common::web3::event_stream::ChainCursor;
 use oprf_core::ddlog_equality::shamir::DLogShareShamir;
-use oprf_test_utils::TEST_ETH_PRIVATE_KEY;
 use oprf_types::crypto::PartyId;
 use oprf_types::service::NodeInformation;
 use oprf_types::{OprfKeyId, ShareEpoch, crypto::OprfPublicKey};
@@ -23,8 +22,8 @@ use sqlx::Row;
 use sqlx::{PgConnection, postgres::PgRow};
 
 async fn postgres_secret_manager() -> eyre::Result<(PostgresDb, &'static str, SanitizedSchema)> {
-    let conn = oprf_test_utils::shared_postgres_testcontainer().await?;
-    let schema = oprf_test_utils::next_test_schema();
+    let conn = nodes_common::test_utils::shared_postgres_testcontainer().await?;
+    let schema = nodes_common::test_utils::next_test_schema();
     let db = postgres_secret_manager_with_schema(conn, schema.clone()).await?;
     Ok((db, conn, schema))
 }
@@ -49,7 +48,9 @@ pub(crate) async fn postgres_secret_manager_with_schema(
 async fn load_node_information_success() -> eyre::Result<()> {
     let (secret_manager, connection_string, schema) = postgres_secret_manager().await?;
 
-    let key = PrivateKeySigner::from_str(TEST_ETH_PRIVATE_KEY)?;
+    let key = PrivateKeySigner::from_str(
+        "0x8b3a350cf5c34c9194ca85829a2df0ec3153be0318b5e2d3348e872092edffba",
+    )?;
     let address = key.address();
     let should_party_id = PartyId(42);
     let should_threshold = NonZeroU16::new(2).expect("2 is non-zero");
@@ -61,7 +62,7 @@ async fn load_node_information_success() -> eyre::Result<()> {
 
     // check that the address is stored in the DB
     let mut pg_connection =
-        oprf_test_utils::open_pg_connection(connection_string, &schema.to_string()).await?;
+        nodes_common::test_utils::open_pg_connection(connection_string, &schema).await?;
     let is_node_information: NodeInformation = sqlx::query_as(
         "SELECT eth_address,party_id,threshold FROM node_information WHERE id = TRUE",
     )
@@ -273,8 +274,7 @@ async fn store_pending_share_without_intermediates_fails() -> eyre::Result<()> {
 #[tokio::test]
 async fn confirm_without_pending_share_fails() -> eyre::Result<()> {
     let (secret_manager, connection_string, schema) = postgres_secret_manager().await?;
-    let mut conn =
-        oprf_test_utils::open_pg_connection(connection_string, &schema.to_string()).await?;
+    let mut conn = nodes_common::test_utils::open_pg_connection(connection_string, &schema).await?;
 
     let oprf_key_id = OprfKeyId::new(U160::from(42));
     let public_key = OprfPublicKey::new(rand::random());
@@ -300,8 +300,7 @@ async fn confirm_without_pending_share_fails() -> eyre::Result<()> {
 #[tokio::test]
 async fn abort_keygen_is_idempotent_and_preserves_confirmed_share() -> eyre::Result<()> {
     let (secret_manager, connection_string, schema) = postgres_secret_manager().await?;
-    let mut conn =
-        oprf_test_utils::open_pg_connection(connection_string, &schema.to_string()).await?;
+    let mut conn = nodes_common::test_utils::open_pg_connection(connection_string, &schema).await?;
 
     let oprf_key_id = OprfKeyId::new(U160::from(42));
     let epoch = ShareEpoch::new(42);
@@ -332,7 +331,7 @@ async fn abort_keygen_is_idempotent_and_preserves_confirmed_share() -> eyre::Res
 async fn store_dlog_share_and_fetch_previous() -> eyre::Result<()> {
     let (secret_manager, connection_string, schema) = postgres_secret_manager().await?;
     let mut pg_connection =
-        oprf_test_utils::open_pg_connection(connection_string, &schema.to_string()).await?;
+        nodes_common::test_utils::open_pg_connection(connection_string, &schema).await?;
 
     let oprf_key_id = OprfKeyId::new(U160::from(42));
     let public_key = OprfPublicKey::new(rand::random());
@@ -459,7 +458,7 @@ async fn store_dlog_share_and_fetch_previous() -> eyre::Result<()> {
 async fn store_dlog_share_as_consumer() -> eyre::Result<()> {
     let (secret_manager, connection_string, schema) = postgres_secret_manager().await?;
     let mut pg_connection =
-        oprf_test_utils::open_pg_connection(connection_string, &schema.to_string()).await?;
+        nodes_common::test_utils::open_pg_connection(connection_string, &schema).await?;
 
     let oprf_key_id = OprfKeyId::new(U160::from(42));
     let public_key = OprfPublicKey::new(rand::random());
@@ -518,7 +517,7 @@ async fn store_dlog_share_as_consumer() -> eyre::Result<()> {
 async fn try_retrieve_random_empty_epochs() -> eyre::Result<()> {
     let (secret_manager, connection_string, schema) = postgres_secret_manager().await?;
     let mut pg_connection =
-        oprf_test_utils::open_pg_connection(connection_string, &schema.to_string()).await?;
+        nodes_common::test_utils::open_pg_connection(connection_string, &schema).await?;
 
     let oprf_key_id = OprfKeyId::new(U160::from(42));
     let public_key = OprfPublicKey::new(rand::random());
@@ -573,8 +572,7 @@ async fn try_retrieve_random_empty_epochs() -> eyre::Result<()> {
 #[tokio::test]
 async fn confirm_after_abort_keygen_fails() -> eyre::Result<()> {
     let (secret_manager, connection_string, schema) = postgres_secret_manager().await?;
-    let mut conn =
-        oprf_test_utils::open_pg_connection(connection_string, &schema.to_string()).await?;
+    let mut conn = nodes_common::test_utils::open_pg_connection(connection_string, &schema).await?;
 
     let oprf_key_id = OprfKeyId::new(U160::from(42));
     let public_key = OprfPublicKey::new(rand::random());
@@ -600,7 +598,7 @@ async fn confirm_after_abort_keygen_fails() -> eyre::Result<()> {
 async fn confirm_same_epoch_without_restaging_is_idempotent() -> eyre::Result<()> {
     let (secret_manager, connection_string, schema) = postgres_secret_manager().await?;
     let mut pg_connection =
-        oprf_test_utils::open_pg_connection(connection_string, &schema.to_string()).await?;
+        nodes_common::test_utils::open_pg_connection(connection_string, &schema).await?;
 
     let oprf_key_id = OprfKeyId::new(U160::from(42));
     let public_key = OprfPublicKey::new(rand::random());
@@ -654,7 +652,7 @@ async fn confirm_same_epoch_without_restaging_is_idempotent() -> eyre::Result<()
 async fn confirm_same_epoch_after_restaging_is_idempotent() -> eyre::Result<()> {
     let (secret_manager, connection_string, schema) = postgres_secret_manager().await?;
     let mut pg_connection =
-        oprf_test_utils::open_pg_connection(connection_string, &schema.to_string()).await?;
+        nodes_common::test_utils::open_pg_connection(connection_string, &schema).await?;
 
     let oprf_key_id = OprfKeyId::new(U160::from(42));
     let public_key = OprfPublicKey::new(rand::random());
@@ -701,8 +699,7 @@ async fn confirm_same_epoch_after_restaging_is_idempotent() -> eyre::Result<()> 
 #[tokio::test]
 async fn delete_oprf_key_material_is_idempotent_and_soft_deletes_share() -> eyre::Result<()> {
     let (secret_manager, connection_string, schema) = postgres_secret_manager().await?;
-    let mut conn =
-        oprf_test_utils::open_pg_connection(connection_string, &schema.to_string()).await?;
+    let mut conn = nodes_common::test_utils::open_pg_connection(connection_string, &schema).await?;
 
     let oprf_key_id = OprfKeyId::new(U160::from(42));
     let public_key = OprfPublicKey::new(rand::random());
@@ -736,8 +733,7 @@ async fn delete_oprf_key_material_is_idempotent_and_soft_deletes_share() -> eyre
 #[tokio::test]
 async fn confirm_deleted_share_returns_store_on_deleted_share() -> eyre::Result<()> {
     let (secret_manager, connection_string, schema) = postgres_secret_manager().await?;
-    let mut conn =
-        oprf_test_utils::open_pg_connection(connection_string, &schema.to_string()).await?;
+    let mut conn = nodes_common::test_utils::open_pg_connection(connection_string, &schema).await?;
 
     let oprf_key_id = OprfKeyId::new(U160::from(42));
     let public_key = OprfPublicKey::new(rand::random());
@@ -770,7 +766,7 @@ async fn confirm_deleted_share_returns_store_on_deleted_share() -> eyre::Result<
 async fn test_delete() -> eyre::Result<()> {
     let (secret_manager, connection_string, schema) = postgres_secret_manager().await?;
     let mut pg_connection =
-        oprf_test_utils::open_pg_connection(connection_string, &schema.to_string()).await?;
+        nodes_common::test_utils::open_pg_connection(connection_string, &schema).await?;
 
     let oprf_key_id = OprfKeyId::new(U160::from(42));
     let public_key = OprfPublicKey::new(rand::random());
